requirements.txt
python-dotenv
html-telegraph-poster
python-telegram-bot
beautifulsoup4